Rename connector_builder_option.py
